get rid of u500vcu118 and u500vc707 devkit directories
