+++ Makefile.vc707-u500devkit
@@ -5,7 +5,7 @@ FPGA_DIR := $(base_dir)/fpga-shells/xilinx
 PROJECT := sifive.fpgashells.shell.xilinx
 MODEL := VC707Shell
 export CONFIG_PROJECT := sifive.freedom.unleashed
-export CONFIG := DevKitFPGAConfig_WithDevKit50MHz
+export CONFIG := DevKitU500VC707FPGAConfig_WithDevKit50MHz
 export BOARD := vc707
 export BOOTROM_DIR := $(base_dir)/bootrom/sdboot
